rename PcfValidator to PcfObserver
