+++ src/PinAssignment/PcfObserver.cpp
@@ -1,4 +1,4 @@
-#include "PcfValidator.h"
+#include "PcfObserver.h"
 #include "Utils/QtUtils.h"
 
 #include <QCryptographicHash>
